+++ src/main/java/oz/yamyam_map/module/member/dto/request/MemberSignupReq.java
@@ -1,4 +1,4 @@
-package oz.yamyam_map.module.member.dto;
+package oz.yamyam_map.module.member.dto.request;
 
 import io.swagger.v3.oas.annotations.media.Schema;
 import jakarta.validation.constraints.NotBlank;
